let proof router handle proof requests
