+++ hierophant/hierophant/src/prover_network_service.rs
@@ -1,6 +1,4 @@
-use crate::hierophant_state::{
-    HierophantState, ProofRequestData, VkHash, WorkerState, WorkerStatus,
-};
+use crate::hierophant_state::{HierophantState, ProofRequestData, VkHash};
 use crate::network::prover_network_server::ProverNetwork;
 use crate::network::{
     CreateProgramRequest, CreateProgramResponse, CreateProgramResponseBody, ExecutionStatus,
@@ -8,6 +6,7 @@ use crate::network::{
     GetProofRequestStatusRequest, GetProofRequestStatusResponse, Program, RequestProofRequest,
     RequestProofResponse, RequestProofResponseBody,
 };
+use crate::proof_router::worker_state::WorkerState;
 use alloy_primitives::{Address, B256};
 use log::{error, info};
 use std::sync::{Arc, Mutex};
@@ -196,90 +195,39 @@ impl ProverNetwork for ProverNetworkService {
         info!("request_proof called");
         let req = request.into_inner();
 
-        // Extract and log the body contents if present
-        if let Some(body) = &req.body {
-            let vk_hash: VkHash = body.vk_hash.into();
-            let vk_hash_hex = vk_hash.to_hex_string();
-            println!("RequestProof request details:");
-            println!("  Nonce: {}", body.nonce);
-            println!("  VK Hash: 0x{}", vk_hash_hex);
-            println!("  Version: {}", body.version);
-            println!("  Mode: {}", body.mode);
-            println!("  Strategy: {}", body.strategy);
-            println!("  Stdin URI: {}", body.stdin_uri);
-            println!("  Deadline: {}", body.deadline);
-            println!("  Cycle Limit: {}", body.cycle_limit);
-            println!("  Gas Limit: {}", body.gas_limit);
-        }
+        let body = match req.body {
+            Some(body) => body,
+            None => {
+                // The client only ever supplies Some(program) [as of sp1-sdk v4.1.3].
+                // It's not clear why it would ever be None.
+                let error_msg = format!(
+                    "No program supplied in body of RequestProof with signature {}",
+                    hex::encode(&req.signature)
+                );
+                error!("{error_msg}");
+                return Err(Status::invalid_argument(error_msg));
+            }
+        };
+
+        // Extract and log the body
+        let vk_hash: VkHash = body.vk_hash.into();
+        let vk_hash_hex = vk_hash.to_hex_string();
+        println!("RequestProof request details:");
+        println!("  Nonce: {}", body.nonce);
+        println!("  VK Hash: 0x{}", vk_hash_hex);
+        println!("  Version: {}", body.version);
+        println!("  Mode: {}", body.mode);
+        println!("  Strategy: {}", body.strategy);
+        println!("  Stdin URI: {}", body.stdin_uri);
+        println!("  Deadline: {}", body.deadline);
+        println!("  Cycle Limit: {}", body.cycle_limit);
+        println!("  Gas Limit: {}", body.gas_limit);
 
         // Log the signature
         println!("Signature: 0x{}", hex::encode(&req.signature));
 
-        // Generate a mock request ID (this would typically be a unique identifier for the proof request)
-        let mut request_id = vec![0u8; 32];
-        let uuid = Uuid::new_v4();
-        let uuid_bytes = uuid.as_bytes();
-        request_id[0..16].copy_from_slice(uuid_bytes);
-
-        // Find an available worker
-        let selected_worker = {
-            let workers = self.state.workers.read().await;
-            workers
-                .values()
-                .find(|w| w.status == WorkerStatus::Idle)
-                .cloned()
-        };
-
-        // If we found a worker, dispatch the job
-        if let Some(worker) = selected_worker {
-            // Spawn a task to send the request to the worker
-            // This runs in the background so we don't block the RPC response
-            let req_id = request_id.clone();
-            tokio::spawn(async move {
-                println!(
-                    "Sending proof request to contemplant {}:{}",
-                    worker.name, worker.id
-                );
-                // // Create the worker task request
-                // let task_request = WorkerTaskRequest {
-                //     request_id: hex::encode(&req_id),
-                //     // Include all necessary parameters from the original request
-                //     vk_hash: hex::encode(&req.body.as_ref().unwrap().vk_hash),
-                //     stdin_uri: req.body.as_ref().unwrap().stdin_uri.clone(),
-                //     // Any other parameters needed...
-                // };
-                //
-                // // Send the request to the worker
-                // match reqwest::Client::new()
-                //     .post(format!("{}/task", worker.address))
-                //     .json(&task_request)
-                //     .send()
-                //     .await
-                // {
-                //     Ok(response) => {
-                //         if response.status().is_success() {
-                //             println!("Task dispatched to worker: {}", worker.id);
-                //             // Update worker status to busy
-                //             // ...
-                //         } else {
-                //             println!("Worker returned error: {}", response.status());
-                //             // Handle error, maybe try another worker
-                //             // ...
-                //         }
-                //     }
-                //     Err(e) => {
-                //         println!("Error sending task to worker: {}", e);
-                //         // Handle error, maybe try another worker
-                //         // ...
-                //     }
-                // }
-            });
-
-            // TODO: implement a request queue.
-        } else {
-            println!("No available workers found for the task");
-            return Err(Status::internal("no worker available"));
-        }
+        // route the proof!!
+        let request_id = self.state.proof_router.route_proof();
 
         // Generate a mock transaction hash
         let mut tx_hash = vec![0u8; 32]; // 32-byte transaction hash
@@ -288,8 +236,9 @@ impl ProverNetwork for ProverNetworkService {
         }
 
         // Store proof request data for status checks
-        let deadline = req.body.as_ref().map(|b| b.deadline).unwrap_or(0);
+        let deadline = body.deadline;
 
+        // TODO: why?
         // Store in our request database
         let proof_request_data = ProofRequestData {
             tx_hash: tx_hash.clone(),
@@ -305,7 +254,7 @@ impl ProverNetwork for ProverNetworkService {
             .proof_requests
             .lock()
             .await
-            .insert(request_id.clone(), proof_request_data);
+            .insert(vk_hash, proof_request_data);
 
         // Create the response
         let response = RequestProofResponse {
@@ -336,62 +285,6 @@ impl ProverNetwork for ProverNetworkService {
             hex::encode(&req.request_id)
         );
 
-        // Look up the request in our database
-        let proof_requests = self.state.proof_requests.lock().await;
-
-        match proof_requests.get(&req.request_id) {
-            Some(proof_data) => {
-                // For testing purposes, let's simulate that the proof is being processed
-                // In a real implementation, you would check the actual status from a database
-
-                // Create the response
-                let response = GetProofRequestStatusResponse {
-                    fulfillment_status: proof_data.fulfillment_status as i32,
-                    execution_status: proof_data.execution_status as i32,
-                    request_tx_hash: proof_data.tx_hash.clone(),
-                    deadline: proof_data.deadline,
-                    fulfill_tx_hash: proof_data.fulfill_tx_hash.clone(),
-                    proof_uri: proof_data.proof_uri.clone(),
-                    public_values_hash: proof_data.public_values_hash.clone(),
-                };
-
-                println!("Responding with status:");
-                println!(
-                    "  Fulfillment Status: {:?}",
-                    FulfillmentStatus::from_i32(response.fulfillment_status).unwrap()
-                );
-                println!(
-                    "  Execution Status: {:?}",
-                    ExecutionStatus::from_i32(response.execution_status).unwrap()
-                );
-                println!(
-                    "  Request Tx Hash: 0x{}",
-                    hex::encode(&response.request_tx_hash)
-                );
-                println!("  Deadline: {}", response.deadline);
-
-                if let Some(ref fulfill_tx_hash) = response.fulfill_tx_hash {
-                    println!("  Fulfill Tx Hash: 0x{}", hex::encode(fulfill_tx_hash));
-                }
-
-                if let Some(ref proof_uri) = response.proof_uri {
-                    println!("  Proof URI: {}", proof_uri);
-                }
-
-                if let Some(ref public_values_hash) = response.public_values_hash {
-                    println!(
-                        "  Public Values Hash: 0x{}",
-                        hex::encode(public_values_hash)
-                    );
-                }
-
-                Ok(Response::new(response))
-            }
-            None => {
-                // If request ID is not found, return an error
-                println!("Request ID not found");
-                Err(Status::not_found("Request ID not found"))
-            }
-        }
+        self.state.proof_router.get_proof_status(&req.request_id)
     }
 }
